bug: fixed error in build by intalling missing devdependencies and refactoring style file
