+++ package.json
@@ -18,6 +18,8 @@
   "devDependencies": {
     "@types/node": "^22.10.2",
     "@types/react": "^19.0.2",
+    "@typescript-eslint/eslint-plugin": "^8.19.1",
+    "@typescript-eslint/parser": "^8.19.1",
     "babel-plugin-styled-components": "^2.1.4",
     "eslint": "^8.0.0",
     "eslint-config-next": "^12.0.0",
