+++ pages/posts/[id].tsx
@@ -4,7 +4,7 @@ import CustomHead from "../../components/CustomHead";
 import { SectionContainer } from "../../styles/SharedStyles";
 import { Post } from "../../lib/types";
 import { getPostById, getPosts } from "../../lib/utils/api";
-import { PostBody } from "./style";
+import { PostBody } from "../../styles/PostStyle";
 
 interface PostProps {
   post: Post;
