+++ styles/PostStyle.ts
